requirements.txt
Flask==3.1.0
qrcode[pil]==8.2
Pillow==11.1.0
Werkzeug==3.1.3
pyzbar==0.1.9